# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Helliox API
  version: 1.0.0

paths:
  /questions:
    get:
      summary: Get a list of questions
      operationId: getQuestions
      responses:
        '200':
          description: A list of question objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /consultation:
    post:
      summary: Submit answers to questions
      operationId: submitConsultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/QuestionAnswer'
      responses:
        '200':
          description: Consultation submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Question:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        question:
          type: string
          example: "What is the capital of France?"
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
      required:
        - uuid
        - question
        - options
    QuestionOption:
      type: object
      properties:
        id:
          type: string
          example: "A"
        option:
          type: string
          example: "Yes"
      required:
        - id
        - option
    QuestionAnswer:
      type: object
      properties:
        questionId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        answer:
          type: String
          example: true
      required:
        - questionId
        - answer
    ConsultationResponse:
      type: object
      properties:
        forbiddenPrescriptions:
          type: array
          items:
            type: string
            example: "Ibuprofen"
      required:
        - forbiddenPrescriptions
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad request"
      required:
        - code
        - message